test: split release helper coverage
