+++ tests/harness/test_blueprint_harness_branches.py
@@ -7,7 +7,6 @@ import tempfile
 import unittest
 
 import scripts.blueprint_harness_branches as branches_mod
-import scripts.blueprint_harness_releases as releases_mod
 
 
 class BlueprintHarnessBranchPolicyTests(unittest.TestCase):
@@ -35,44 +34,6 @@ class BlueprintHarnessBranchPolicyTests(unittest.TestCase):
             self.assertEqual(policy.default_dev_branch, "v4.29.0")
             self.assertEqual(policy.required_backport_branches, ())
             self.assertEqual(policy.source_path, root / "branch-policy.json")
-
-    def test_release_candidate_names_normalize_to_tags_and_branch_ids(self) -> None:
-        self.assertEqual(releases_mod.normalize_release_candidate_name("4.30-rc2"), "4.30-rc2")
-        self.assertEqual(releases_mod.normalize_release_candidate_name("v4.30.0-rc2"), "4.30-rc2")
-        self.assertEqual(releases_mod.release_candidate_name_or_none("v4.30.0-rc2"), "4.30-rc2")
-        self.assertIsNone(releases_mod.release_candidate_name_or_none("v4.30.0"))
-        self.assertEqual(releases_mod.release_candidate_ref("4.30-rc2"), "v4.30.0-rc2")
-        self.assertEqual(releases_mod.normalize_lean_release_ref("4.30-rc2"), "v4.30.0-rc2")
-        self.assertEqual(releases_mod.release_branch_from_lean_ref("leanprover/lean4:v4.30.0-rc2"), "v4.30.0")
-
-    def test_lean_release_order_key_orders_release_candidates_before_final_release(self) -> None:
-        self.assertLess(
-            releases_mod.lean_release_order_key("v4.30.0-rc1"),
-            releases_mod.lean_release_order_key("v4.30.0-rc2"),
-        )
-        self.assertLess(
-            releases_mod.lean_release_order_key("v4.30.0-rc2"),
-            releases_mod.lean_release_order_key("v4.30.0"),
-        )
-        self.assertEqual(
-            releases_mod.lean_release_order_key("v4.30-rc2"),
-            releases_mod.lean_release_order_key("v4.30.0-rc2"),
-        )
-        self.assertIsNone(releases_mod.lean_release_order_key("nightly-testing"))
-
-    def test_rewrite_lean_toolchain_preserves_existing_final_newline_style(self) -> None:
-        with tempfile.TemporaryDirectory() as tmp:
-            root = Path(tmp)
-            with_newline = root / "with-newline" / "lean-toolchain"
-            without_newline = root / "without-newline" / "lean-toolchain"
-            self.write(with_newline, "leanprover/lean4:v4.29.0\n")
-            self.write(without_newline, "leanprover/lean4:v4.29.0")
-
-            releases_mod.rewrite_lean_toolchain(with_newline, "v4.30.0")
-            releases_mod.rewrite_lean_toolchain(without_newline, "v4.30.0")
-
-            self.assertEqual(with_newline.read_text(encoding="utf-8"), "leanprover/lean4:v4.30.0\n")
-            self.assertEqual(without_newline.read_text(encoding="utf-8"), "leanprover/lean4:v4.30.0")
 
     def test_active_release_branch_uses_release_id_for_release_candidate_toolchain(self) -> None:
         with tempfile.TemporaryDirectory() as tmp:
